@@ -3,11 +3,12 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using Microsoft.OpenApi.Models;
 using NBitcoin;
 using Swashbuckle.AspNetCore.Swagger;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using WalletWasabi.Backend.Middlewares;
@@ -35,15 +36,21 @@ namespace WalletWasabi.Backend
 			services.AddMvc(options => options.ModelMetadataDetailsProviders.Add(new SuppressChildValidationMetadataProvider(typeof(BitcoinAddress))))
 				.AddControllersAsServices();
 
+			services.AddMvc()
+				.AddNewtonsoftJson();
+
+			services.AddControllers()
+				.AddNewtonsoftJson();
+
 			// Register the Swagger generator, defining one or more Swagger documents
 			services.AddSwaggerGen(c =>
 			{
-				c.SwaggerDoc($"v{Constants.BackendMajorVersion}", new Info
+				c.SwaggerDoc($"v{Constants.BackendMajorVersion}", new OpenApiInfo
 				{
 					Version = $"v{Constants.BackendMajorVersion}",
 					Title = "Wasabi Wallet API",
 					Description = "Privacy focused, ZeroLink compliant Bitcoin Web API.",
-					License = new License { Name = "Use under MIT.", Url = "https://github.com/zkSNACKs/WalletWasabi/blob/master/LICENSE.md" }
+					License = new OpenApiLicense { Name = "Use under MIT.", Url = new Uri("https://github.com/zkSNACKs/WalletWasabi/blob/master/LICENSE.md") }
 				});
 
 				// Set the comments path for the Swagger JSON and UI.
@@ -63,7 +70,7 @@ namespace WalletWasabi.Backend
 		// This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
 #pragma warning disable IDE0060 // Remove unused parameter
 
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env, Global global)
+		public void Configure(IApplicationBuilder app, IWebHostEnvironment env, Global global)
 #pragma warning restore IDE0060 // Remove unused parameter
 		{
 			app.UseStaticFiles();
@@ -74,6 +81,8 @@ namespace WalletWasabi.Backend
 			// Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.), specifying the Swagger JSON endpoint.
 			app.UseSwaggerUI(c => c.SwaggerEndpoint($"/swagger/v{Constants.BackendMajorVersion}/swagger.json", "Wasabi Wallet API V3"));
 
+			app.UseRouting();
+
 			// So to correctly handle HEAD requests.
 			// https://www.tpeczek.com/2017/10/exploring-head-method-behavior-in.html
 			// https://github.com/tpeczek/Demo.AspNetCore.Mvc.CosmosDB/blob/master/Demo.AspNetCore.Mvc.CosmosDB/Middlewares/HeadMethodMiddleware.cs
@@ -81,9 +90,9 @@ namespace WalletWasabi.Backend
 
 			app.UseResponseCompression();
 
-			app.UseMvc();
+			app.UseEndpoints(endpoints => endpoints.MapControllers());
 
-			var applicationLifetime = app.ApplicationServices.GetRequiredService<IApplicationLifetime>();
+			var applicationLifetime = app.ApplicationServices.GetRequiredService<IHostApplicationLifetime>();
 			applicationLifetime.ApplicationStopping.Register(() => OnShutdown(global)); // Don't register async, that won't hold up the shutdown
 		}
 
