@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -14,10 +15,10 @@ namespace WalletWasabi.Backend
 			where T : class, IStartupTask
 			=> services.AddTransient<IStartupTask, T>();
 
-		public static async Task RunWithTasksAsync(this IWebHost webHost, CancellationToken cancellationToken = default)
+		public static async Task RunWithTasksAsync(this IHost host, CancellationToken cancellationToken = default)
 		{
-			await Task.WhenAll(webHost.Services.GetServices<IStartupTask>().Select(t => t.ExecuteAsync(cancellationToken)));
-			await webHost.RunAsync(cancellationToken);
+			await Task.WhenAll(host.Services.GetServices<IStartupTask>().Select(t => t.ExecuteAsync(cancellationToken)));
+			await host.RunAsync(cancellationToken);
 		}
 	}
 }
