@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 using NBitcoin;
 using NBitcoin.RPC;
 using System;
@@ -23,10 +24,12 @@ namespace WalletWasabi.Backend
 			{
 				var endPoint = "http://localhost:37127/";
 
-				using var host = WebHost.CreateDefaultBuilder(args)
-					.UseStartup<Startup>()
-					.UseUrls(endPoint)
+				using var host = Host.CreateDefaultBuilder(args)
+					.ConfigureWebHostDefaults(webBuilder => webBuilder
+							.UseStartup<Startup>()
+							.UseUrls(endPoint))
 					.Build();
+
 				await host.RunWithTasksAsync();
 			}
 			catch (Exception ex)
